refactor: keep only MongoDB ObjectID rule
