+++ internal/messagetemplates/templates.go
@@ -67,7 +67,6 @@ const (
 	StringTimeZoneTemplate
 	StringKubernetesQualifiedNameTemplate
 	StringMongoDBObjectIDTemplate
-	StringMongoDBConnectionStringTemplate
 	URLTemplate
 	SliceUniqueTemplate
 	UniquePropertiesTemplate
@@ -185,8 +184,7 @@ var rawMessageTemplates = map[templateKey]string{
 	string must be a valid Kubernetes Qualified Name
 {{- end }}
 `,
-	StringMongoDBObjectIDTemplate:         "string must be a 24-character lowercase hexadecimal MongoDB ObjectID",
-	StringMongoDBConnectionStringTemplate: "string must be a valid MongoDB connection string{{- if .Error }}: {{ .Error }}{{- end }}",
+	StringMongoDBObjectIDTemplate: "string must be a 24-character lowercase hexadecimal MongoDB ObjectID",
 	SliceUniqueTemplate: `elements are not unique, {{ .Custom.FirstOrdinal }} and {{ .Custom.SecondOrdinal }} elements collide
 {{- if gt (len .Custom.Constraints) 0 }} based on constraints: {{ joinSlice .Custom.Constraints "" }}{{- end }}`,
 	UniquePropertiesTemplate: `all of [{{ joinSlice .ComparisonValue "" }}] properties must be unique, but '{{ .Custom.FirstProperty }}' collides with '{{ .Custom.SecondProperty }}'
